tests/wpt/tests/gyroscope/Gyroscope-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that gyroscope is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://w3c.github.io/sensors/#permissions-policy-api">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('gyroscope', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise gyroscope.');
</script>
